# src/MarketingBundle/Controller/CampaignEventController.php
<?php

namespace MarketingBundle\Controller;

use FOS\RestBundle\Controller\Annotations\Route;
use FOS\RestBundle\View\View;
use Knp\Bundle\PaginatorBundle\Pagination\SlidingPagination;
use Mullenlowe\CommonBundle\Controller\MullenloweRestController;
use Symfony\Component\HttpFoundation\Request;
use FOS\RestBundle\Controller\Annotations as Rest;
use MarketingBundle\Enum\PaginateEnum;
use Swagger\Annotations as SWG;

/**
 * Class CampaignEventController
 * @package MarketingBundle\Controller
 * @Route("campaign-event")
 */
class CampaignEventController extends MullenloweRestController
{
    const CONTEXT = 'CampaignEvent';

    /**
     * @Rest\Get("/{id}", requirements={"id"="\d+"})
     * @Rest\View()
     *
     * @SWG\Get(
     *     path="/campaign-event/{id}",
     *     summary="Get CampaignEvent",
     *     operationId="getCampaignEvent",
     *     tags={"CampaignEvent"},
     *     @SWG\Parameter(
     *         name="id",
     *         in="query",
     *         type="integer",
     *         required=false,
     *         description="CampaignEvent ID"
     *     ),
     *     @SWG\Response(
     *         response="200",
     *         description="A CampaignEvent",
     *         @SWG\Definition(ref="#/definitions/CampaignEventContext")
     *     ),
     *     @SWG\Response(
     *         response=404,
     *         description="not found",
     *         @SWG\Schema(ref="#/definitions/Error")
     *     )
     * )
     *
     * @param int $id
     * @return View
     */
    public function getAction(int $id)
    {
        $campaignEvent = $this->getDoctrine()
            ->getRepository('MarketingBundle:CampaignEvent')
            ->find($id);

        if (empty($campaignEvent)) {
            throw $this->createNotFoundException('Compain Event not found');
        }

        return $this->createView($campaignEvent);
    }

    /**
     * @Rest\Get("/")
     * @Rest\View()
     *
     * @SWG\Get(
     *     path="/campaign-event",
     *     summary="Get CampaignEvents",
     *     operationId="getCampaignEvents",
     *     tags={"CampaignEvent"},
     *     @SWG\Parameter(
     *         name="eventTypeId",
     *         in="query",
     *         type="integer",
     *         required=false,
     *         description="Find campaign events by eventTypeId"
     *     ),
     *     @SWG\Parameter(
     *         name="limit",
     *         in="query",
     *         type="integer",
     *         required=false,
     *         description="limit"
     *     ),
     *     @SWG\Response(
     *         response="200",
     *         description="CampaignEvents",
     *         @SWG\Definition(ref="#/definitions/CampaignEventContextMulti")
     *     ),
     *     @SWG\Response(
     *         response=404,
     *         description="not found",
     *         @SWG\Schema(ref="#/definitions/Error")
     *     )
     * )
     *
     * @param Request $request
     * @return View
     */
    public function cgetAction(Request $request)
    {
        $repository = $this->getDoctrine()->getRepository('MarketingBundle:CampaignEvent');

        if (!empty($request->query->getInt('limit'))
            && $request->query->getInt('limit') === -1
            && $request->query->getInt('eventTypeId')
        ) {
            $queryBuilder = $repository->findBy(['eventType' => $request->query->getInt('eventTypeId')]);

            return $this->createView($queryBuilder);
        }

        $paginator = $this->get('knp_paginator');
        /** @var SlidingPagination $pager */
        $pager = $paginator->paginate(
            $repository->createQueryBuilder('campaignEvent'),
            $request->query->getInt('page', PaginateEnum::CURRENT_PAGE),
            $request->query->getInt('limit', PaginateEnum::LIMIT)
        );

        return $this->createPaginatedView($pager);
    }
}