# web/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Marketing Api",
        "version": "1.0.0"
    },
    "host": "api5.audi.dev.agence-one.net",
    "basePath": "/marketing/dev/",
    "schemes": [
        "http"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/campaign-event/{id}": {
            "get": {
                "tags": [
                    "CampaignEvent"
                ],
                "summary": "Get CampaignEvent",
                "operationId": "getCampaignEvent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "CampaignEvent ID",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A CampaignEvent",
                        "schema": {
                            "$ref": "#/definitions/CampaignEventContext"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/campaign-event": {
            "get": {
                "tags": [
                    "CampaignEvent"
                ],
                "summary": "Get CampaignEvents",
                "operationId": "getCampaignEvents",
                "parameters": [
                    {
                        "name": "eventTypeId",
                        "in": "query",
                        "description": "Find campaign events by eventTypeId",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CampaignEvents",
                        "schema": {
                            "$ref": "#/definitions/CampaignEventContextMulti"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Context": {
            "properties": {
                "context": {
                    "type": "string"
                }
            },
            "type": "object"
        },
...
        "Error": {
            "required": [
                "errors"
            ],
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "properties": {
                            "message": {
                                "type": "string"
                            },
                            "code": {
                                "type": "integer"
                            },
                            "type": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "type": "object",
            "allOf": [
                {
                    "$ref": "#/definitions/Context"
                }
            ]
        },
...
        "CampaignEventContext": {
            "properties": {
                "data": {
                    "type": "object",
                    "allOf": [
                        {
                            "$ref": "#definitions/CampaignEvent"
                        }
                    ]
                }
            },
            "allOf": [
                {
                    "$ref": "#definitions/Context"
                }
            ]
        },
        "CampaignEventContextMulti": {
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#definitions/CampaignEvent"
                    }
                }
            },
            "allOf": [
                {
                    "$ref": "#definitions/Context"
                }
            ]
        },
...
    }
}